# backend/openapi.internal.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaaS Boilerplate",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auth/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Access Token",
        "description": "Returns a Token, or - when the user has two-factor auth enabled - an\nMfaChallengeOut to exchange at POST /auth/mfa/verify.",
        "operationId": "get_access_token_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_get_access_token_v1_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Token"
                    },
                    {
                      "$ref": "#/components/schemas/MfaChallengeOut"
                    }
                  ],
                  "title": "Response Get Access Token V1 Auth Token Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/api-tokens": {
      "get": {
        "tags": [
          "API Tokens"
        ],
        "summary": "List Api Tokens",
        "operationId": "list_api_tokens_v1_api_tokens_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiTokenResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Tokens V1 Api Tokens Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Tokens"
        ],
        "summary": "Create Api Token",
        "operationId": "create_api_token_v1_api_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/api-tokens/{token_id}": {
      "delete": {
        "tags": [
          "API Tokens"
        ],
        "summary": "Revoke Api Token",
        "operationId": "revoke_api_token_v1_api_tokens__token_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiTokenCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Permissions"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "permissions"
        ],
        "title": "ApiTokenCreate"
      },
      "ApiTokenCreatedResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "token_prefix": {
            "type": "string",
            "title": "Token Prefix"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "is_expired": {
            "type": "boolean",
            "title": "Is Expired",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "token_prefix",
          "permissions",
          "created_at",
          "expires_at",
          "last_used_at",
          "revoked_at",
          "token",
          "is_expired"
        ],
        "title": "ApiTokenCreatedResponse"
      },
      "ApiTokenResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "token_prefix": {
            "type": "string",
            "title": "Token Prefix"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "is_expired": {
            "type": "boolean",
            "title": "Is Expired",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "token_prefix",
          "permissions",
          "created_at",
          "expires_at",
          "last_used_at",
          "revoked_at",
          "is_expired"
        ],
        "title": "ApiTokenResponse"
      },
...
      "Body_get_access_token_v1_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_get_access_token_v1_auth_token_post"
      },
...
      "MfaChallengeOut": {
        "properties": {
          "mfa_required": {
            "type": "boolean",
            "const": true,
            "title": "Mfa Required",
            "default": true
          },
          "mfa_token": {
            "type": "string",
            "title": "Mfa Token"
          }
        },
        "type": "object",
        "required": [
          "mfa_required",
          "mfa_token"
        ],
        "title": "MfaChallengeOut",
        "description": "Returned instead of a Token when the password was correct but a second\nfactor is required. Exchange it at POST /auth/mfa/verify."
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "token_type"
        ],
        "title": "Token"
      },
...
      "ErrorResponse": {
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "time",
          "path",
          "method",
          "error_code",
          "msg"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ValidationDetail": {
        "properties": {
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "field": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Field",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "ctx": {
            "additionalProperties": true,
            "title": "Ctx",
            "type": "object"
          }
        },
        "required": [
          "error_code",
          "field",
          "msg",
          "ctx"
        ],
        "title": "ValidationDetail",
        "type": "object"
      },
      "ValidationErrorResponse": {
        "$defs": {
          "ValidationDetail": {
            "properties": {
              "error_code": {
                "title": "Error Code",
                "type": "string"
              },
              "field": {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "title": "Field",
                "type": "array"
              },
              "msg": {
                "title": "Msg",
                "type": "string"
              },
              "ctx": {
                "additionalProperties": true,
                "title": "Ctx",
                "type": "object"
              }
            },
            "required": [
              "error_code",
              "field",
              "msg",
              "ctx"
            ],
            "title": "ValidationDetail",
            "type": "object"
          }
        },
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "time",
          "path",
          "method",
          "error_code",
          "msg",
          "errors"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  }
}